[Fusili] Remove UID and associated functionality (#1918)
We have converged on a different mechanism for ensuring MLIR value names
are unique that doesn't rely on UIDs. So right now the only need for
this seems to be for holding a map from tensor UID to the host/dev
pointers to the `g->execute()` call. However, we could also just use the
tensor name or TensorAttr directly. Removing UID tracking to keep things
simple - this relieves us of having to maintain two mechanisms of
uniqueness checking - UID & name.