[Fusilli] Refactor stride utilities to support ambiguous (unit dim) cases (#2283)
We can't always calculate the stride order from stride, especially in
ambiguous (unit dim) cases. For example a tensor of shape `(256, 128, 1,
1)` would have contiguous stride `(128, 1, 1, 1)` but we can't certainly
say the stride order for this is `(3, 2, 1, 0)` as it could also be `(3,
0, 1, 2)`.

A more robust way of getting the stride order (hence layout) is to
reconstruct the stride from the dims and compare it with the known
stride order for contiguous vs channels last.

Apart from this change, the PR also removes the contiguity check from
TensorAttr (letting it be more relaxed) and moves it to the pre/post
validate checks on the ConvFPropNode itself. This makes sense because
certain layout restrictions are op-semantics driven and we should model
those in the op specific validations rather than on the generic
TensorAttr.

Tests updated accordingly.