[Fusilli] Generalize physical tensor layout support (#2653)
Currently, Fusilli only supports 2 physical tensor layouts: NCHW
(contiguous) and NHWC. However, pointwise ops and matmul ops
(https://github.com/nod-ai/shark-ai/pull/2632) require support for any
layout that can be represented as a permutation of dimensions.

## Changes

#### 1. Added new validation method `hasValidPhysicalRepresentation()`
This method checks that the combination of stride & logical dimensions
can be represented as a physical tensor and that the logical tensor is a
transpose of the physical tensor.

#### 2. Generalized `getPhysicalDim()`
This now supports getting the physical dims for when they are a
permutation of the logical dims. This uses `getLogicalToPhysicalPerm()`
to get the physical dims by transposing the logical dims.

#### 3. Updated the ASM emitters
Updated the ASM emitters to use the new helper
`getLogicalToPhysicalPerm()`. This is the same utility used by
`getPhysicalDims()`, which ensures that the transpose vec matches the
expected type. This is mostly an NFC and provides a cleaner and more
concise way to get the permutation.

Also, pointwise nodes weren't converting the physical tensor to a
logical tensor, which was incorrect. This PR breaks pointwise ops
without fixing this, so the fix for pointwise ops is included here.

#### 4. Updated CHECK lines in conv tests
`getPhysicalDim` assumed that physical memory would be stored only in
NCHW or NHWC format. That meant there was a unique way to go from
logical to physical. Now that this is generalized, we run into a
problem: unit dims add ambiguity and so there is no longer a unique
physical tensor representation of the logical dims. Unit dims can be
placed anywhere without changing the meaning of the tensor. However, we
still need a single unique representation for the MLIR types to match.
When finding the physical tensor, this PR chooses to place the non-unit
dims in the same location as the logical tensor. This provides a way to
find a unique physical representation of the tensor.


For example:

```
k = 256, c = 128, r = 1, s = 1
dim = [k, c, r, s] = [256, 128, 1, 1]
stride = [c * r * s, r * s, s, 1] = [128, 1, 1, 1]
```

Previously, this would result in a physical tensor of
`!torch.vtensor<[256,1,1,128],f32>` but now this results in
`!torch.vtensor<[256,128,1,1],f32>`.

---------

Signed-off-by: Ian Wood <ianwood@u.northwestern.edu>
Co-authored-by: Sambhav Jain <sambhav.eee@gmail.com>